+++ src/MacVim/MMWindowController.h
@@ -51,12 +51,12 @@
 - (void)setTitle:(NSString *)title;
 - (void)setDocumentFilename:(NSString *)filename;
 - (void)setToolbar:(NSToolbar *)toolbar;
-- (void)createScrollbarWithIdentifier:(long)ident type:(int)type;
-- (BOOL)destroyScrollbarWithIdentifier:(long)ident;
-- (BOOL)showScrollbarWithIdentifier:(long)ident state:(BOOL)visible;
-- (void)setScrollbarPosition:(int)pos length:(int)len identifier:(long)ident;
+- (void)createScrollbarWithIdentifier:(int32_t)ident type:(int)type;
+- (BOOL)destroyScrollbarWithIdentifier:(int32_t)ident;
+- (BOOL)showScrollbarWithIdentifier:(int32_t)ident state:(BOOL)visible;
+- (void)setScrollbarPosition:(int)pos length:(int)len identifier:(int32_t)ident;
 - (void)setScrollbarThumbValue:(float)val proportion:(float)prop
-                    identifier:(long)ident;
+                    identifier:(int32_t)ident;
 - (void)setDefaultColorsBackground:(NSColor *)back foreground:(NSColor *)fore;
 - (void)setFont:(NSFont *)font;
 - (void)setWideFont:(NSFont *)font;
